add has_slot method; refactor python to create a slot in allocate_slots if a slot does not exist for the request
