+++ lib/bindings/python/src/dynamo/llm/vllm_integration/kv_cache_manager.py
@@ -43,8 +43,19 @@ class KvbmCacheManager:
         """
         Get the computed blocks for the request.
         """
+        sequence_hashes = self._create_slot(request)
+
+        owned_blocks = self.cache_manager.get_computed_blocks(sequence_hashes)
+        block_count = owned_blocks.block_count()
+
+        return KvbmCacheBlocks(owned_blocks), block_count
+
+    def _create_slot(self, request: Request) -> list[int]:
+        """Create a slot for the request."""
         if bool(request.mm_positions):
             raise ValueError("Unsupported request - requires mm extra keys")
+
+        all_token_ids = request.all_token_ids
 
         # extract the critial aspects of the request that effect how the tokens are hashed
         request = KvbmRequest(
@@ -55,14 +66,7 @@ class KvbmCacheManager:
             salt_hash=request.cache_salt,
         )
 
-        # todo(vllm): determine if this call should be idempotent or if it should fail
-        # if the slot already exists
-        sequence_hashes = self.cache_manager.create_slot(request, request.all_token_ids)
-
-        owned_blocks = self.cache_manager.get_computed_blocks(sequence_hashes)
-        block_count = owned_blocks.block_count()
-
-        return KvbmCacheBlocks(owned_blocks), block_count
+        return self.cache_manager.create_slot(request, all_token_ids)
 
     def allocate_slots(
         self,
@@ -111,9 +115,12 @@ class KvbmCacheManager:
         if num_new_tokens == 0:
             raise ValueError("num_new_tokens must be greater than 0")
 
+        if not self.cache_manager.has_slot(request.request_id):
+            self._create_slot(request)
+
         num_computed_tokens = request.num_computed_tokens + num_new_computed_tokens
 
-        # we need to extract from the request the new tokens to append to to the block state
+        # we need to extract from the request the new tokens to append to the block state
         prev_computed_tokens = self.cache_manager.num_computed_tokens(
             request.request_id
         )
@@ -124,11 +131,14 @@ class KvbmCacheManager:
         slot_update = SlotUpdate(
             request_id=request.request_id,
             request_num_tokens=request.num_tokens,
+            request_num_computed_tokens=request.num_computed_tokens,
             tokens_to_append=tokens_to_append,
             num_new_tokens=num_new_tokens,
             num_new_computed_tokens=num_new_computed_tokens,
             new_computed_blocks=new_computed_blocks,
-            num_lookahead_blocks=num_lookahead_tokens,
+            # TODO(ryan): add support for lookahead blocks
+            # comment out for now, otherwise would error out
+            # num_lookahead_blocks=num_lookahead_tokens,
             delay_cache_blocks=delay_cache_blocks,
         )
 
@@ -138,7 +148,7 @@ class KvbmCacheManager:
             return None
 
         new_blocks = [
-            KVCacheBlock(block_id=block.block_id) for block in new_blocks.block_ids()
+            KVCacheBlock(block_id=block_id) for block_id in new_blocks.block_ids()
         ]
 
         return KVCacheBlocks(blocks=new_blocks)
